+++ src/main/java/seedu/address/model/contact/Contact.java
@@ -12,8 +12,8 @@ import seedu.address.model.item.NamedItem;
 import seedu.address.model.item.TaggedItem;
 import seedu.address.model.item.commons.Name;
 import seedu.address.model.item.commons.Tag;
-import seedu.address.ui.ContactCard;
-import seedu.address.ui.DisplayableItem;
+import seedu.address.ui.card.ContactCard;
+import seedu.address.ui.util.DisplayableItem;
 import seedu.address.ui.UiPart;
 
 /**
