+++ src/main/java/seedu/address/model/todo/Todo.java
@@ -16,8 +16,8 @@ import seedu.address.model.item.commons.Datetime;
 import seedu.address.model.item.commons.Location;
 import seedu.address.model.item.commons.Name;
 import seedu.address.model.item.commons.Tag;
-import seedu.address.ui.DisplayableItem;
-import seedu.address.ui.TodoCard;
+import seedu.address.ui.util.DisplayableItem;
+import seedu.address.ui.card.TodoCard;
 import seedu.address.ui.UiPart;
 
 /**
