+++ src/main/java/seedu/address/ui/ListPanel.java
@@ -8,6 +8,7 @@ import javafx.scene.control.ListCell;
 import javafx.scene.control.ListView;
 import javafx.scene.layout.Region;
 import seedu.address.commons.core.LogsCenter;
+import seedu.address.ui.util.DisplayableItem;
 
 /**
  * Panel containing the unified list that can display contacts, events, or todos.
