+++ src/main/java/seedu/address/ui/card/ContactCard.java
@@ -1,4 +1,4 @@
-package seedu.address.ui;
+package seedu.address.ui.card;
 
 import java.util.Comparator;
 
@@ -8,6 +8,7 @@ import javafx.scene.layout.FlowPane;
 import javafx.scene.layout.HBox;
 import javafx.scene.layout.Region;
 import seedu.address.model.contact.Contact;
+import seedu.address.ui.UiPart;
 
 /**
  * An UI component that displays information of a {@code Contact}.
