+++ src/main/java/seedu/address/ui/card/EventCardFactory.java
@@ -1,6 +1,7 @@
-package seedu.address.ui;
+package seedu.address.ui.card;
 
 import seedu.address.model.event.Event;
+import seedu.address.ui.UiPart;
 
 public class EventCardFactory implements CardFactory<Event> {
     @Override
